Improve implementation and test
There is a bug when an entry originated with a smaller ID alters an
entry with bigger ID, the tests were updated in order to reveal it.

The solution is to use the clock time as the transaction ID and keep it
when consolidating the model in ledger.cpp. Independently of the order
of processing, the most recent version of replaces is kept.

We still need to test the case of concurrent edits.

Signed-off-by: Aeliton G. Silva <aeliton@aeliton.com>