Fix clock comparison issues
Because clocks are partial ordered (some of them are not comparable),
the regular < operator from std::map won't give the behaviour we want.

The solution so far is to have a pure comparison function that will
return true only if the secont argument is comparable and greater than
the first argument. This is Journal::smaller(a, b).

Signed-off-by: Aeliton G. Silva <aeliton@aeliton.com>