Move shared code from Broker to EntryHandler
It is possible to share more code as Journals and Brokers ought to share
the same API.

The differences in behaviour now are:
- Broker keep a list of data providers whilst Journals are their single
  provider
- Journals record the entries on insert whilst Brokers just shared them

We have now the distance of the source of data being shared. This will
be useful when the journals reconnect on a different broker.

During attach() the EntryHandler says where it will be listening from
inserts on the attaching EntryHandler, this will be useful together with
the distance to update the preferred incoming port of a data source
(shortest distance wins, for example).

The connection context infra-structure now is used to hold the journal
clock version. The first context is reserved for the Entry handler own
use. To insert data on itself the index 0 (default on insert) has to be
used.

Signed-off-by: Aeliton G. Silva <aeliton@aeliton.com>