Centralize connect functions
As connections are handled by the same function, there was still a non
public function in the Broker API that used a BrokerStub as argument.
This function has been removed in favor of the public API function.

The ConnectionData struct has been removed from the public interface.
This is to remove duplication of BrokerStub objects. Now the connection
is the single holder of the stub objects and the cached data.

Signed-off-by: Aeliton G. Silva <aeliton@aeliton.com>